implemented Terminal endpoint
